+++ src/GenerateFFMpegCommand.py
@@ -2,6 +2,7 @@
 class FFMpegCommand:
     def __init__(self,
                  video_encoder: str,
+                 video_encoder_speed: str,
                  video_quality: str,
                  video_pixel_format: str,
                  audio_encoder: str,
@@ -11,6 +12,7 @@ class FFMpegCommand:
                  color_primaries: str,
                  color_transfer: str,):
         self._video_encoder = video_encoder
+        self._video_encoder_speed = video_encoder_speed
         self._video_quality = video_quality
         self._video_pixel_format = video_pixel_format
         self._audio_encoder = audio_encoder
@@ -58,6 +60,19 @@ class FFMpegCommand:
                         command +=["-crf","23"]
                     case "Low":
                         command +=["-crf","28"]
+                
+                match self._video_encoder_speed:
+                    case "placebo":
+                        command +=["-preset","placebo"]
+                    case "slow":
+                        command +=["-preset","slow"]
+                    case "medium":
+                        command +=["-preset","medium"]
+                    case "fast":
+                        command +=["-preset","fast"]
+                    case "fastest":
+                        command +=["-preset","veryfast"]
+
                 if self._hdr_mode:
                     command += ["-x264-params", f'"{encoder_params}"']
                 
@@ -74,6 +89,18 @@ class FFMpegCommand:
                     case "Low":
                         command +=["-crf","28"]
 
+                match self._video_encoder_speed:
+                    case "placebo":
+                        command +=["-preset","placebo"]
+                    case "slow":
+                        command +=["-preset","slow"]
+                    case "medium":
+                        command +=["-preset","medium"]
+                    case "fast":
+                        command +=["-preset","fast"]
+                    case "fastest":
+                        command +=["-preset","veryfast"]
+                        
                 if self._hdr_mode:
                     command += ["-x265-params", f'"{encoder_params}"']
                 
@@ -88,6 +115,19 @@ class FFMpegCommand:
                         command +=["-crf","30"]
                     case "Low":
                         command +=["-crf","40"]
+
+                match self._video_encoder_speed:
+                    case "placebo":
+                        command +=["-preset","placebo"]
+                    case "slow":
+                        command +=["-preset","slow"]
+                    case "medium":
+                        command +=["-preset","medium"]
+                    case "fast":
+                        command +=["-preset","fast"]
+                    case "fastest":
+                        command +=["-preset","veryfast"]
+
             case "av1":
                 command +=["-c:v","libsvtav1"]
                 match self._video_quality:
@@ -99,6 +139,17 @@ class FFMpegCommand:
                         command +=["-cq:v","23"]
                     case "Low":
                         command +=["-cq:v","28"]
+                match self._video_encoder_speed:
+                    case "placebo":
+                        command +=["-preset","0"]
+                    case "slow":
+                        command +=["-preset","4"]
+                    case "medium":
+                        command +=["-preset","8"]
+                    case "fast":
+                        command +=["-preset","12"]
+                    case "fastest":
+                        command +=["-preset","13"]
             case "ffv1":
                 command +=["-c:v","ffv1"]
                 match self._video_quality:
@@ -138,8 +189,21 @@ class FFMpegCommand:
                         command +=["-cq:v","23"]
                     case "Low":
                         command +=["-cq:v","28"]
+
                 """if self._hdr_mode:
                     command += ["-x264-params", f'"{encoder_params}"']"""
+                match self._video_encoder_speed:
+                    case "placebo":
+                        command +=["-preset","p7"]
+                    case "slow":
+                        command +=["-preset","p6"]
+                    case "medium":
+                        command +=["-preset","p4"]
+                    case "fast":
+                        command +=["-preset","p2"]
+                    case "fastest":
+                        command +=["-preset","p1"]
+
             case "x265_nvenc":
                 command +=["-c:v","hevc_nvenc"]
                 match self._video_quality:
@@ -153,6 +217,17 @@ class FFMpegCommand:
                         command +=["-cq:v","28"]
                 """if self._hdr_mode:
                     command += ["-x265-params", f'"{encoder_params}"']"""
+                match self._video_encoder_speed:
+                    case "placebo":
+                        command +=["-preset","p7"]
+                    case "slow":
+                        command +=["-preset","p6"]
+                    case "medium":
+                        command +=["-preset","p4"]
+                    case "fast":
+                        command +=["-preset","p2"]
+                    case "fastest":
+                        command +=["-preset","p1"]
             case "av1_nvenc":
                 command +=["-c:v","av1_nvenc"]
                 match self._video_quality:
@@ -164,6 +239,18 @@ class FFMpegCommand:
                         command +=["-cq:v","23"]
                     case "Low":
                         command +=["-cq:v","28"]
+                match self._video_encoder_speed:
+                    case "placebo":
+                        command +=["-preset","p7"]
+                    case "slow":
+                        command +=["-preset","p6"]
+                    case "medium":
+                        command +=["-preset","p4"]
+                    case "fast":
+                        command +=["-preset","p2"]
+                    case "fastest":
+                        command +=["-preset","p1"]
+                
             case "h264_vaapi":
                 command +=["-c:v","h264_vaapi"]
                 match self._video_quality: